Implement new inheritance hierarchie in KatanaSwipe
Reduces code duplication.
