+++ WASD/Assets/Scripts/WeaponAmmunition/KatanaSwipes/KatanaSwipe.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9378bf846a0db974db5ff327f5a0bbcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
